# api/openapi.yaml
openapi: 3.0.3
info:
  title: "Sport Shop"
  description: This is an API reference for my sport shop pet project based on OAS 3.0 specification. Check out my other projects at [https://github.com/kirich-yo](https://github.com/kirich-yo).
  contact:
    email: gdleader03@gmail.com
  license:
    name: GNU Lesser General Public License v3.0
    url: https://www.gnu.org/licenses/lgpl-3.0.html
  version: 1.0.0
servers:
  - url: https://development.restapi-sportshop.com/api/v1
    description: Development server
  - url: https://staging.restapi-sportshop.com/api/v1
    description: Staging server
  - url: https://api.restapi-sportshop.com/api/v1
    description: Production server
...
paths:
  /item:
    get:
      tags:
        - item
      summary: Get all items from the catalog.
      description: Returns all items from the catalog.
      operationId: getItems
      responses:
        '200':
          description: A list of items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
            application/xml:
              schema:
                type: array
                xml:
                  name: items
                  wrapped: true
                items:
                  $ref: '#/components/schemas/Item'
        '500':
          description: Internal server error.
    post:
      tags:
        - item
      summary: Add an item to the catalog
      description: Adds an item.
      operationId: addItem
      requestBody:
        $ref: '#/components/requestBodies/Item'
      responses:
        '201':
          description: Item created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
            application/xml:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Bad request body.
        '401':
          description: The client is not authorized.
        '403':
          description: The client does not have appropriate permissions to perform the operation.
        '500':
          description: Internal server error.
      security:
        - sportshop_jwt: []
  /item/{itemID}:
    get:
      tags:
        - item
      summary: Get an item by ID
      description: Returns the item by ID.
      operationId: getItem
      parameters:
        - name: itemID
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: A single item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
            application/xml:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid ID supplied.
        '404':
          description: Item not found.
        '500':
          description: Internal server error.
    patch:
      tags:
        - item
      summary: Update item info
      description: Updates the item by ID.
      operationId: updateItem
      parameters:
        - name: itemID
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        $ref: '#/components/requestBodies/Item'
      responses:
        '200':
          description: The item has been succesfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
            application/xml:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid ID supplied.
        '401':
          description: The client is not authorized.
        '403':
          description: The client does not have appropriate permissions to perform the operation.
        '404':
          description: Item not found.
        '500':
          description: Internal server error.
      security:
        - sportshop_jwt: []
    delete:
      tags:
        - item
      summary: Delete an item from the catalog
      description: Deletes the item by ID.
      operationId: deleteItem
      parameters:
        - name: itemID
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: The item has been succesfully deleted.
        '400':
          description: Invalid ID supplied.
        '401':
          description: The client is not authorized.
        '403':
          description: The client does not have appropriate permissions to perform the operation.
        '500':
          description: Internal server error.
      security:
        - sportshop_jwt: []
...
  /user/{username}:
    get:
      tags:
        - user
      summary: Get user account info
      description: Returns user info by login.
      operationId: getUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied.
        '404':
          description: User not found.
    patch:
      tags:
        - user
      summary: Change user account info
      description: Changes account info.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        $ref: '#/components/requestBodies/User'
      responses:
        '200':
          description: Account info has been changed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid username supplied.
        '401':
          description: The client is not authorized.
        '403':
          description: The client does not have appropriate permissions to perform the operation.
        '404':
          description: User not found.
      security:
        - sportshop_jwt: []
    delete:
      tags:
        - user
      summary: Delete user account
      description: Deletes a user account.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: The user has been successfully deleted.
        '400':
          description: Invalid username supplied.
        '401':
          description: The client is not authorized.
        '403':
          description: The client does not have appropriate permissions to perform the operation.
      security:
        - sportshop_jwt: []
  /user/{username}/role:
    get:
      tags:
        - user
      summary: Get user role
      description: Returns user role by username.
      operationId: getUserRole
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
            application/xml:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Invalid username supplied.
        '404':
          description: User not found.
...
  /role/{roleID}:
    get:
      tags:
        - role
      summary: Get a role by ID
      description: Returns a role by ID.
      operationId: getRole
      parameters:
        - name: roleID
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: A role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
            application/xml:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Invalid ID supplied.
        '404':
          description: Role not found.
...
components:
  schemas:
    Role:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: Customer
        permissions:
          type: array
          items:
            type: string
      xml:
        name: role
...
    User:
      type: object
      required:
        - id
        - firstName
        - phoneNumber
        - role
      properties:
        id:
          type: integer
          format: int32
          example: 456
        username:
          type: string
          example: johndoe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        dateOfBirth:
          type: string
          format: date-time
          example: 2000-01-01
        photoURL:
          type: string
          example: https://cdn.sportshop.com/59b1f1ce-7299-4e9b-93c4-cb5b94641864.jpg
      xml:
        name: user
    Item:
      type: object
      required:
        - id
        - name
        - price
        - photoURL
      properties:
        id:
          type: integer
          format: int32
          example: 123
        name:
          type: string
          example: Dumbbells Set of 2 Hex Rubber Encased Dumbbells, Weights Dumbbells Set with Metal Handle for Exercise and Fitness (25 lb, Pair)
        categoryID:
          type: integer
          format: int32
          example: 1
        price:
          type: number
          format: float
          example: 111.82
        salePrice:
          type: number
          format: float
          example: 95.047
        rating:
          type: number
          format: float
          minimum: 1.00
          maximum: 5.00
          example: 4.76
        reviewCount:
          type: integer
          format: int32
          example: 5794
        photoURL:
          type: string
          example: https://cdn.sportshop.com/032e216c-5630-44a5-b98e-0c77f76a8df4.jpg
      xml:
        name: item
...
  requestBodies:
    Item:
      description: An item.
      required: true
      content:
        application/json:
          schema:
            properties:
              name:
                type: string
                example: Dumbbells Set of 2 Hex Rubber Encased Dumbbells, Weights Dumbbells Set with Metal Handle for Exercise and Fitness (25 lb, Pair)
              categoryID:
                type: integer
                format: int32
                example: 1
              price:
                type: number
                format: float
                example: 111.82
              salePrice:
                type: number
                format: float
                example: 95.047
              photoURL:
                type: string
                example: https://cdn.sportshop.com/032e216c-5630-44a5-b98e-0c77f76a8df4.jpg
        application/xml:
          schema:
            properties:
              name:
                type: string
                example: Dumbbells Set of 2 Hex Rubber Encased Dumbbells, Weights Dumbbells Set with Metal Handle for Exercise and Fitness (25 lb, Pair)
              categoryID:
                type: integer
                format: int32
                example: 1
              price:
                type: number
                format: float
                example: 111.82
              salePrice:
                type: number
                format: float
                example: 95.047
              photoURL:
                type: string
                example: https://cdn.sportshop.com/032e216c-5630-44a5-b98e-0c77f76a8df4.jpg
            xml:
              name: item
    User:
      description: A user.
      required: true
      content:
        application/json:
          schema:
            properties:
              username:
                type: string
                example: johndoe
              firstName:
                type: string
                example: John
              lastName:
                type: string
                example: Doe
              dateOfBirth:
                type: string
                format: date-time
                example: 2000-01-01
              photoURL:
                type: string
                example: https://cdn.sportshop.com/59b1f1ce-7299-4e9b-93c4-cb5b94641864.jpg
              password:
                type: string
                format: password
                example: "********"
        application/xml:
          schema:
            properties:
              firstName:
                type: string
                example: John
              lastName:
                type: string
                example: Doe
              dateOfBirth:
                type: string
                format: date-time
                example: 2000-01-01
              photoURL:
                type: string
                example: https://cdn.sportshop.com/59b1f1ce-7299-4e9b-93c4-cb5b94641864.jpg
              password:
                type: string
                format: password
                example: "********"
            xml:
              name: user
...
  securitySchemes:
    sportshop_jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT